+++ migrations/2026-01-17-074444_season-10-part-2/down.sql
@@ -1,2 +1,5 @@
+alter table data.player_versions
+    drop column priority;
+
 alter table data.player_report_versions
     alter column quote set not null;