+++ migrations/2026-01-17-074444_season-10-part-2/up.sql
@@ -1,6 +1,9 @@
 alter table data.player_report_versions
     alter column quote drop not null;
 
+alter table data.player_versions
+    add column priority float8; -- null = this player version is from before priority was available
+
 -- TODO:
 --   - Priority, within BaseAttributes
 --   - Pitch types and their probabilities, within BaseAttributes