+++ mmoldb-db/src/schema/data_schema.rs
@@ -439,6 +439,7 @@ pub mod data {
             num_modifications -> Int4,
             occupied_equipment_slots -> Array<Nullable<Text>>,
             included_report_categories -> Array<Nullable<Int8>>,
+            priority -> Nullable<Float8>,
         }
     }